rework listing screen to a more finished level

+++ app/src/main/java/eu/kutscheid/elegoomonitor/data/model/PrinterData.kt
@@ -34,9 +34,9 @@ data class PrintInfo(
     @SerialName("Status")
     val status: Int,
     @SerialName("CurrentLayer")
-    val currentLayer: Int,
+    val currentLayer: Long,
     @SerialName("TotalLayer")
-    val totalLayer: Int,
+    val totalLayer: Long,
     @SerialName("CurrentTicks")
     val currentTicks: Long,
     @SerialName("TotalTicks")
